+++ package.json
@@ -14,6 +14,7 @@
     "@tabler/icons-react": "^2.30.0",
     "antd": "^5.8.2",
     "axios": "0.27.2",
+    "diacriticless": "^1.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "8.0.1",
