+++ src/component/CheckIn/CheckIn.css
@@ -1,0 +1,88 @@
+.page-checkin{
+    background-image: url('../../assets/BackRoud.png');
+    height: auto;
+    background-size: cover;
+    display: flex;
+    padding-bottom: 3%;
+  
+}
+.main-checkin{
+    margin: 3% 17% 0 17%;
+    width: 100%;
+    background-color: rgb(255, 255, 255,0.70);
+    box-shadow: 0px 10px 32px #006885;
+    backdrop-filter: blur(3px);
+  
+    border-radius: 5px;
+}
+.form-search-checkin{
+    width: 100%;
+    padding: 40px 20px 20px 20px;
+    border-right: 1px solid #006885;
+    .title{
+        font-size: 20px;
+        font-weight: 600;
+        padding-bottom: 20px;
+    }
+    .notification{
+        font-size: 16px;
+        font-weight: 500;
+        padding-bottom: 20px;
+    }
+}
+.btn-checkin{
+    margin-left: 20%;
+    width: 150px;
+    height:40px;
+    font-size: 18px;
+    font-weight: 600;
+   background-color: #006885;
+   color: white;
+   &:hover{
+    color: white !important;
+    background-color: #006885;
+    border: none !important;
+    }
+}
+.lable-form{
+    font-size: 18px !important;
+}
+.input-form{
+    height: 40px;
+}
+.notification-checkin{
+    padding: 15px ;
+}
+:where(.css-dev-only-do-not-override-17a39f8).ant-input:hover {
+border-color:none !important;
+
+}
+.notification-header{
+ font-size: 18px;
+}
+.notification-title{
+    padding-top: 10px;
+    font-size: 18px;
+    font-weight: 600 !important;
+    color: red;
+}
+.notification-item-small{
+    padding-top: 5px;
+    font-size: 17px;
+    font-weight: 600 !important;
+    
+}
+.notification-ul{
+    font-size: 17px;
+    padding-left: 20px;
+    padding-top: 5px;
+}
+.notification{
+    font-size: 17px;
+    padding-bottom: 20px;
+    padding-top: 20px;
+}
+.notification-ul-li{
+    padding-left: 30px;
+    padding-top: 5px;
+}
