+++ src/component/SearchBookingMy/SearchBookingMy.css
@@ -1,0 +1,40 @@
+.notification-checkin{
+    background-image: url('../../assets/SGN-SearchBooking.webp');
+    background-size: cover;
+    width: 100%;
+
+  
+}
+.page-search-booking{
+    background-image: url('../../assets/BackRoud.png');
+    background-size: cover;
+    display: flex;
+    width: 100%;
+    height: 100vh;
+
+
+}
+.main-search-booking{
+    margin: 7% 17% 0 17%;
+    width: 100%;
+    background-color: rgb(255, 255, 255,0.70);
+    box-shadow: 0px 10px 32px #006885;
+    backdrop-filter: blur(3px);
+    border-radius: 5px;
+    height: 500px;
+
+}
+.btn-search-booking{
+    margin-left: 30%;
+    width: 150px;
+    height:40px;
+    font-size: 18px;
+    font-weight: 600;
+   background-color: #006885;
+   color: white;
+   &:hover{
+    color: white !important;
+    background-color: #006885;
+    border: none !important;
+    }
+}
