+++ src/component/SelectFlight/SelectFlight.jsx
@@ -1,27 +1,26 @@
-import { Row, Col, Form, Button, Radio, Collapse, Input, DatePicker, Select } from 'antd';
-import '../Home/Passengers.css'
+import { Row, Col, Form, Button } from 'antd';
+import '../SelectFlight/SelectFlight.css'
 import { IconPlane, IconUserCircle, IconCurrencyDollar, IconShoppingCart } from '@tabler/icons-react';
-import { CaretRightOutlined } from '@ant-design/icons';
-import { useNavigate, useLocation } from 'react-router-dom';
+import imgVeluxe from '../../assets/wdelux.svg'
+import imgEco from '../../assets/weco.svg'
+import imgskyboss from '../../assets/wskyboss.svg'
+import imgBusinesswhite from '../../assets/businesswhite.svg'
+import imgNoflight from '../../assets/noflight.svg'
+
+import { useNavigate } from 'react-router-dom';
+import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
-import { useState } from 'react';
 
-const Passengers = () => {
+
+const SelectFlight = () => {
     const navigate = useNavigate();
-    const [value, setValue] = useState(1);
     const data_homepage = useSelector(state => state.homepage.data_booking);
+
     const numberBooking = data_homepage.adult + data_homepage.children;
-
-    const onChange = (e) => {
-        setValue(e.target.value);
-    };
-    const numberadult = Array.from({ length: data_homepage.adult });
-    const numberChildren = Array.from({ length: data_homepage.children });
-    const numberbaby = Array.from({ length: data_homepage.baby });
-
 
     return (
         <div className="select-flight">
+
             <div className="info-flight">
                 <Row>
                     <Col span={16} className='infor-select'>
@@ -46,7 +45,7 @@ const Passengers = () => {
                     <Col span={8} className='icon-selcet'>
                         <Row >
                             <IconPlane style={{ color: '#006885', width: 30, height: 30, marginRight: 15 }} />
-                            <IconUserCircle style={{ color: '#006885', width: 30, height: 30, marginRight: 15 }} />
+                            <IconUserCircle style={{ color: 'grey', width: 30, height: 30, marginRight: 15 }} />
                             <IconShoppingCart style={{ color: 'grey', width: 30, height: 30, marginRight: 15 }} />
                             <IconCurrencyDollar style={{ color: 'grey', width: 30, height: 30, marginRight: 15 }} />
                         </Row>
@@ -58,234 +57,98 @@ const Passengers = () => {
             <div className='mains-container'>
                 <Row>
                     <Col span={15} className='infor-user-select-flight'>
-                        {numberadult.map((_, index) => (
-                            <div key={index}>
-                                <Collapse
-                                    size="large"
-                                    items={[
-                                        {
-                                            key: '1',
-                                            label: <div style={{ fontSize: '18px', fontWeight: 600 }}>Người lớn</div>,
-                                            children:
-                                                <div className='formPassengers'>
-                                                    <Row className='rowInforPassengers'>
-                                                        <Radio.Group onChange={onChange} value={value}>
-                                                            <Radio value='Nam'>Nam</Radio>
-                                                            <Radio value='Nữ'>Nữ</Radio>
-                                                            <Radio value='Khác'>Khác</Radio>
-                                                        </Radio.Group>
-                                                    </Row>
-                                                    <Row className='rowInforPassengers'>
-                                                        <Col span={12}>
-                                                            <span style={{ fontSize: '16px', fontWeight: 500 }}>Họ*</span>
-                                                            <Input placeholder='Họ' style={{ width: 350 }} />
-                                                        </Col>
-                                                        <Col span={12}>
-                                                            <span style={{ fontSize: '16px', fontWeight: 500 }}>Tên đệm & tên*</span>
-                                                            <Input placeholder='Tên đệm & tên theo trình tự CCCD' />
-                                                        </Col>
-                                                    </Row>
-                                                    <Row className='rowInforPassengers'>
-                                                        <Col span={12}>
-                                                            <span style={{ fontSize: '16px', fontWeight: 500 }}>Ngày sinh*</span>
-                                                            <DatePicker style={{ width: 350 }} placeholder='Ngày sinh' />
-                                                        </Col>
-                                                        <Col span={12}>
-                                                            <span style={{ fontSize: '16px', fontWeight: 500 }}>Quốc gia*</span>
-                                                            <Select
-                                                                showSearch
-                                                                style={{ width: 365 }}
-                                                                placeholder="Điểm khởi hành"
-                                                                optionFilterProp="children"
-                                                                filterOption={(input, option) => (option?.label ?? '').includes(input)}
-                                                                filterSort={(optionA, optionB) =>
-                                                                    (optionA?.label ?? '').toLowerCase().localeCompare((optionB?.label ?? '').toLowerCase())
-                                                                }
+                        <Row className='icon-class'>
+                            <Col span={7} >
 
-                                                                options={[
-                                                                    {
-                                                                        value: '1',
-                                                                        label: 'SGN',
-                                                                    },
-                                                                    {
-                                                                        value: '2',
-                                                                        label: 'HN',
-                                                                    },
+                            </Col>
+                            <Col span={4} className='imgBusinesswhite'>
+                                <img src={imgBusinesswhite} style={{ width: 110, height: 70, paddingLeft: 20 }} />
+                            </Col>
+                            <Col span={4} className='imgskyboss'>
+                                <img src={imgskyboss} style={{ width: 110, height: 70, paddingLeft: 20 }} />
+                            </Col>
+                            <Col span={4} className='imgVeluxe'>
+                                <img src={imgVeluxe} style={{ width: 110, height: 70, paddingLeft: 20 }} />
+                            </Col>
+                            <Col span={4} className='imgEco'>
+                                <img src={imgEco} style={{ width: 110, height: 70, paddingLeft: 20 }} />
+                            </Col>
+                        </Row>
+                        <Row className='icon-class'>
+                            <Col span={7} className='symbol-airline'>
+                                <Row>
+                                    <Col span={24}>
+                                        <span className='code-airline'>VJ198</span>
+                                    </Col>
+                                    <Col span={24}>
+                                        <Row>
+                                            <Col span={9}>
+                                                <span className='clock-airline'>05:25</span>
+                                            </Col>
+                                            <Col span={6}>
+                                                <span className='code-airline'>Đến</span>
+                                            </Col>
+                                            <Col span={9}>
+                                                <span className='clock-airline'>07:35</span>
+                                            </Col>
+                                        </Row>
+                                    </Col>
+                                    <Col span={24}>
+                                        <span className='code-airline'>Airbus A321</span>
 
-                                                                ]}
-                                                            />
-                                                        </Col>
-                                                    </Row>
-                                                    <Row className='rowInforPassengers'>
-                                                        <Col span={12}>
-                                                            <span style={{ fontSize: '16px', fontWeight: 500 }}>Số điện thoại*</span>
-                                                            <Input placeholder='Số điện thoại' style={{ width: 350 }} />
-                                                        </Col>
-                                                        <Col span={12}>
-                                                            <span style={{ fontSize: '16px', fontWeight: 500 }}>Email*</span>
-                                                            <Input placeholder='Email của bạn' />
-                                                        </Col>
-                                                    </Row>
-                                                    <Row className='rowInforPassengers'>
-                                                        <Col span={24}>
-                                                            <span style={{ fontSize: '16px', fontWeight: 500 }}>Nơi ở</span>
-                                                            <Input placeholder='Nơi ở hiện tại' style={{ width: 730 }} />
-                                                        </Col>
-                                                    </Row>
-                                                </div>
-                                        },
-                                    ]}
-                                    expandIconPosition='end'
-                                    expandIcon={({ isActive }) => <CaretRightOutlined rotate={isActive ? 90 : 0} />}
-                                    style={{
-                                        backgroundColor: '#F1F1F1',
-                                        marginTop: '10px'
-                                    }}
-                                /></div>
-                        ))}
-                        {data_homepage.children == 0 ? ''
-                            :
-                            <div>
-                                {numberChildren.map((_, index) => (
-                                    <Collapse
-                                        size="large"
-                                        items={[
-                                            {
-                                                key: '1',
-                                                label: <div style={{ fontSize: '18px', fontWeight: 600 }}>Trẻ em</div>,
-                                                children:
-                                                    <div className='formPassengers'>
-                                                        <Row className='rowInforPassengers'>
-                                                            <Col span={12}>
-                                                                <span style={{ fontSize: '16px', fontWeight: 500 }}>Họ*</span>
-                                                                <Input placeholder='Họ' style={{ width: 350 }} />
-                                                            </Col>
-                                                            <Col span={12}>
-                                                                <span style={{ fontSize: '16px', fontWeight: 500 }}>Tên đệm & tên*</span>
-                                                                <Input placeholder='Tên đệm & tên theo trình tự theo giấy khai sinh' />
-                                                            </Col>
-                                                        </Row>
-                                                        <Row className='rowInforPassengers'>
-                                                            <Col span={12}>
-                                                                <span style={{ fontSize: '16px', fontWeight: 500 }}>Ngày sinh*</span>
-                                                                <DatePicker style={{ width: 350 }} placeholder='Ngày sinh' />
-                                                            </Col>
-                                                            <Col span={12}>
-                                                                <span style={{ fontSize: '16px', fontWeight: 500 }}>Giới tính</span>
-                                                                <Select
-                                                                    showSearch
-                                                                    style={{ width: 365 }}
-                                                                    defaultValue={1}
-                                                                    optionFilterProp="children"
-                                                                    filterOption={(input, option) => (option?.label ?? '').includes(input)}
-                                                                    filterSort={(optionA, optionB) =>
-                                                                        (optionA?.label ?? '').toLowerCase().localeCompare((optionB?.label ?? '').toLowerCase())
-                                                                    }
+                                    </Col>
+                                </Row>
 
-                                                                    options={[
-                                                                        {
-                                                                            value: '1',
-                                                                            label: 'Nam',
-                                                                        },
-                                                                        {
-                                                                            value: '2',
-                                                                            label: 'Nữ',
-                                                                        },
-
-                                                                    ]}
-                                                                />
-                                                            </Col>
-                                                        </Row>
-
-                                                    </div>
-                                            },
-                                        ]}
-                                        expandIconPosition='end'
-                                        expandIcon={({ isActive }) => <CaretRightOutlined rotate={isActive ? 90 : 0} />}
-                                        style={{
-                                            backgroundColor: '#F1F1F1',
-                                            marginTop: '10px'
-                                        }}
-                                    />
-                                ))
-                                }
-                            </div>
-                        }
-
-                        {data_homepage.baby == 0 ? ''
-                            :
-                            <div>
-                                {numberbaby.map((_, index) => (
-                                    <Collapse
-                                        size="large"
-                                        items={[
-                                            {
-                                                key: '1',
-                                                label: <div style={{ fontSize: '18px', fontWeight: 600 }}>Em bé</div>,
-                                                children:
-                                                    <div className='formPassengers'>
-                                                        <Row className='rowInforPassengers'>
-                                                            <Col span={24}>
-                                                                <span style={{ fontSize: '16px', fontWeight: 500 }}>Đi cùng</span>
-                                                                <Input placeholder='Tên người em bé đi cùng' style={{ width: 730 }} />
-                                                            </Col>
-                                                        </Row>
-                                                        <Row className='rowInforPassengers'>
-                                                            <Col span={12}>
-                                                                <span style={{ fontSize: '16px', fontWeight: 500 }}>Họ*</span>
-                                                                <Input placeholder='Họ' style={{ width: 350 }} />
-                                                            </Col>
-                                                            <Col span={12}>
-                                                                <span style={{ fontSize: '16px', fontWeight: 500 }}>Tên đệm & tên*</span>
-                                                                <Input placeholder='Tên đệm & tên theo trình tự theo giấy khai sinh' />
-                                                            </Col>
-                                                        </Row>
-                                                        <Row className='rowInforPassengers'>
-                                                            <Col span={12}>
-                                                                <span style={{ fontSize: '16px', fontWeight: 500 }}>Ngày sinh*</span>
-                                                                <DatePicker style={{ width: 350 }} placeholder='Ngày sinh' />
-                                                            </Col>
-                                                            <Col span={12}>
-                                                                <span style={{ fontSize: '16px', fontWeight: 500 }}>Giới tính</span>
-                                                                <Select
-                                                                    showSearch
-                                                                    style={{ width: 365 }}
-                                                                    defaultValue={1}
-                                                                    optionFilterProp="children"
-                                                                    filterOption={(input, option) => (option?.label ?? '').includes(input)}
-                                                                    filterSort={(optionA, optionB) =>
-                                                                        (optionA?.label ?? '').toLowerCase().localeCompare((optionB?.label ?? '').toLowerCase())
-                                                                    }
-
-                                                                    options={[
-                                                                        {
-                                                                            value: '1',
-                                                                            label: 'Nam',
-                                                                        },
-                                                                        {
-                                                                            value: '2',
-                                                                            label: 'Nữ',
-                                                                        },
-
-                                                                    ]}
-                                                                />
-                                                            </Col>
-                                                        </Row>
-
-                                                    </div>
-                                            },
-                                        ]}
-                                        expandIconPosition='end'
-                                        expandIcon={({ isActive }) => <CaretRightOutlined rotate={isActive ? 90 : 0} />}
-                                        style={{
-                                            backgroundColor: '#F1F1F1',
-                                            marginTop: '10px'
-                                        }}
-                                    />
-                                ))
-                                }
-                            </div>
-                        }
+                            </Col>
+                            <Col span={4} className='price-airline'>
+                                <Row>
+                                    <Col span={24}>
+                                        <span className='price'><img src={imgNoflight} /></span>
+                                    </Col>
+                                </Row>
+                                <Row>
+                                    <Col span={24}>
+                                        <span className='price-vnd'>Hết chỗ</span>
+                                    </Col>
+                                </Row>
+                            </Col>
+                            <Col span={4} className='price-airline'>
+                                <Row>
+                                    <Col span={24}>
+                                        <span className='price'><i>1,250</i></span>
+                                    </Col>
+                                </Row>
+                                <Row>
+                                    <Col span={24}>
+                                        <span className='price-vnd'><i>000 VND</i></span>
+                                    </Col>
+                                </Row>
+                            </Col>
+                            <Col span={4} className='price-airline'>
+                                <Row>
+                                    <Col span={24}>
+                                        <span className='price'><i>1,000</i></span>
+                                    </Col>
+                                </Row>
+                                <Row>
+                                    <Col span={24}>
+                                        <span className='price-vnd'><i>000 VND</i></span>
+                                    </Col>
+                                </Row>
+                            </Col>
+                            <Col span={4} className='price-airline'>
+                                <Row>
+                                    <Col span={24}>
+                                        <span className='price'><i>900</i></span>
+                                    </Col>
+                                </Row>
+                                <Row>
+                                    <Col span={24}>
+                                        <span className='price-vnd'><i>000 VND</i></span>
+                                    </Col>
+                                </Row>
+                            </Col>
+                        </Row>
 
                     </Col>
                     <Col span={9} >
@@ -535,8 +398,7 @@ const Passengers = () => {
             <div className="footer">
                 <Row>
                     <Col span={6}>
-                        <Button className='footer-back'
-                            onClick={() => { navigate('/select-fight-infor') }} >Quay lại</Button>
+
                     </Col>
                     <Col span={12} >
                         <Row>
@@ -547,7 +409,7 @@ const Passengers = () => {
 
                     </Col>
                     <Col span={6}>
-                        <Button className='footer-continue' onClick={() => navigate('/select-service')} >Tiếp tục</Button>
+                        <Button className='footer-continue' onClick={() => { navigate('/select-fight-infor') }} >Tiếp tục</Button>
                     </Col>
                 </Row>
             </div>
@@ -556,4 +418,4 @@ const Passengers = () => {
         </div >
     )
 }
-export default Passengers;
+export default SelectFlight;
