+++ src/redux/action/PassengersAction.js
@@ -1,0 +1,7 @@
+export const DATA_PASSENGERS = 'DATA_PASSENGERS';
+export const Data_Passengers = (data) => {
+    return {
+        type: DATA_PASSENGERS,
+        payload: data
+    }
+}
